Strict unpadding
Whenever unpadding a piece of data where the padding was voluntarily
broken or corrupted, the resulting data would be corrupted rather than
crashing.

For a binary `<<1,2,3>>`, wrongfully padded as `<<1,2,3,3,3,2>>`,
The unpadded version would be returned as `<<1,2,3,3>>`.

This patch fixes it so the unpadding strictly follows the padding rule,
and for the example above, the error `bad_padding` will be raised. This
maintains the existing API for correct data, just narrows it for
incorrect ones.